requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.11.0
django-countries==7.2.1
django-country2==0.0.8
django-filter==21.1
django-heroku==0.3.1
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.1.0
factory-boy==3.2.1
Faker==11.3.0
frozenlist==1.2.0
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
maxminddb==2.2.0
multidict==5.2.0
psycopg2==2.9.3
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
urllib3==1.26.7
whitenoise==5.3.0






